using same prompt

+++ examples/llm/tech_qa.py
@@ -237,12 +237,13 @@ def train(args, data_lists):
     gnn = GAT(in_channels=768, hidden_channels=hidden_channels,
               out_channels=1024, num_layers=num_gnn_layers, heads=4)
     # freeze LM
-    #llm = LLM(model_name=args.llm_generator_name).eval()
-    llm = LLM(model_name=args.llm_generator_name, dtype=torch.float32).eval()
+    llm = LLM(model_name=args.llm_generator_name).eval()
+    #
     return llm, DataLoader(
         data_lists["train"] + data_lists["validation"] + data_lists["test"],
         batch_size=eval_batch_size, drop_last=False, pin_memory=True,
         shuffle=False)
+    #
     model = GRetriever(llm=llm, gnn=gnn)
     save_name = "tech-qa-model.pt"
     if os.path.exists(save_name):
